+++ ExamplePlugin/Forms/SuperForm/Items.cs
@@ -31,7 +31,7 @@ namespace HedgehogUtils.Forms.SuperForm
             emeraldTier.pickupRules = ItemTierDef.PickupRules.Default;
             emeraldTier.name = HedgehogUtilsPlugin.Prefix + "_CHAOS_EMERALD_TIER";
             emeraldTier.highlightPrefab = Addressables.LoadAssetAsync<GameObject>(RoR2BepInExPack.GameAssetPaths.Version_1_39_0.RoR2_Base_UI.HighlightTier1Item_prefab).WaitForCompletion();
-            emeraldTier.dropletDisplayPrefab = Addressables.LoadAssetAsync<GameObject>(RoR2BepInExPack.GameAssetPaths.Version_1_39_0.RoR2_Base_Common.Tier1Orb_prefab).WaitForCompletion();
+            emeraldTier.dropletDisplayPrefab = Assets.chaosEmeraldDropletPrefab;
             emeraldTier.canScrap = false;
             emeraldTier.colorIndex = ColorCatalog.ColorIndex.Tier1Item;
             emeraldTier.darkColorIndex = ColorCatalog.ColorIndex.Tier1ItemDark;